Huge redesign of Block-Scheme interaction
Con Man blocks are noew just arbitrary RTT::TaskContexts with the
conman::HookService loaded on them, named "conman"

We now assume someone who wants to interact with a conman block will get
a pointer to the HookService, so we eliminated the subclassing and the
acquiring of operation callers.

Since blocks are no longer subclassed, the user now registers the
execution hooks with the HookService.

The clarifications have been made on port classifications:
- only input ports can have exclusion properties
- only output ports can be designated with a layer

The api now reflects these classifications. Instead of registering a
port with conman along with all of its properties simultaneously, the
act of setting a property makes conman "aware" of that port, and
incorporates it into the scheduling. This is also much cleaner and
simplifies the amount that the user needs to express.

A few other changes, prefer TaskContext pointers as keys instead of
strings, etc, etc.